For pull requests, add the target branch repo to the dependencies
When building a pull request, the name of the target branch tells us
what is the baseline we are targeting.  When fetching dependencies, we
want to make sure that the corresponding repo is enabled.

For example, if a pull request is made against the xenial_-_qt59 branch,
regardless of the branch name used in the user repo we want to fetch the
dependencies from xenial_-_qt59 as well.