Add push notification sending.
This adds a call to braodcast a push notification when a new image
is promoted to the channel for a device. We use '0' as the revision
number, to avoid the need to parse it out, as system-image on the
client side handles this to mean whatever the latest version is in
the channel. This also fixes some of the more extraneous flake8 issues
in the code for the promote-images script.