build-source.sh: support multiple upstream tarball
build-source.sh will read the ubports.source_location 2 lines at a time,
and stop if no more line is available. This way, the format of the file
is (mostly) preserved. The only change is that, by using shell script's
read command, the ubports.source_location must ends with a newline
charactor.

This is required for building gst-droid, as it requires the tarball for
itself and the tarball for gstreamer's "common" repo.