build-binary: fix debsigning for multidist case
Debsign is done at the end of the script, assuming that changes files
are at the root. However, that isn't the case for multidist building,
where packages for each distro are built in its directory.

This commit move the original code to the non-multidist case. Then, in
multidist case, make sure to iterate over all distros and sign changes
files in that directory before archiving them.