build-source: move ubports.* files before actually building source
For non-native packages, if we don't move them away dpkg-buildpackage
will complain that they're local changes and refuse to create a source
package. So, move (instead of copy) them away earlier.

Note that some package already workaround this problem by also including
them as a patch. The method should still work; the file should not be
there in the patch application process anyway.