build-repo: do not allow aptly publication failure anymore
In preparation for the migration to Aptly repository in OTA-15, we have
to make sure that all packages can be published to Aptly repo. This
commit removes the code that ignores failures from Aptly so that it
fails the build, bringing that package to our attention.

For a build to be publishable on Aptly, all files listed in the .dsc
file must be available. This includes .ddeb's (debugging symbols)
and, for some package, .udeb's (packages for Debian Installer), which
was not archived with the old Jenkinsfile. Thus, a package needs to
have its Jenkinsfile updated to be publishable on Aptly.

Fixes https://github.com/ubports/ubuntu-touch/issues/1567