build-source: Add support for debian/watch and depricate ubports.source_location
This adds support for using debian/watch files as the method to download
sources. This is a better way as it will automatically find and download
the version defined by changelog. This is also the official debian way
of doing this, this simplifies our merge with salsa debian folders. This
also adds support for checking for newer upstream versions if needed.

This also prints a deprecation warning when using source location lower
then 3.0 and when using ubports.source_location