build-binary: fix finding multiple orig tarballs
Turns out, I forgot that the format for additional orig tarball is
".orig-<part name>.tar.*", not ".orig.<part name>.tar.*". Thus, the
find rule doesn't work.

This commit fixes the find rule so that it correctly includes the part
tarball as well.

While we're at it, fix the comment too.